--- backend_rentelektro/envs/mongodb.example.env
@@ -1,7 +1,0 @@
-MONGO_USER=
-MONGO_PASSWORD=
-MONGO_HOST=mongodb
-MONGO_PORT=27017
-MONGO_DB_NAME=
-
-CLIENT_SECRET=
