+++ backend_rentelektro/main.py
@@ -2,15 +2,21 @@ import os
 
 from fastapi import FastAPI
 
-from api.api_v1.endpoints import users, tools, rentals, reviews
-from api.api_v1.endpoints import maintance
-from core import models
+from users import endpoints as users
+from tools import endpoints as tools
+from reviews import endpoints as reviews
+from rentals import endpoints as rentals
+from maintance import endpoints as maintance
+from users import models as users_db
+from tools import models as tools_db
+from reviews import models as reviews_db
+from rentals import models as rentals_db
 from core.database import engine
 from prometheus_fastapi_instrumentator import Instrumentator
 from fastapi.middleware.cors import CORSMiddleware
 from debug_toolbar.middleware import DebugToolbarMiddleware
 
-app = FastAPI(tags="Main", root_path="/api")
+app = FastAPI(tags="RentElektro", root_path="/api")
 app.debug = os.getenv("DEBUG", "true").lower() == "true"
 
 app.add_middleware(
@@ -21,8 +27,12 @@ app.add_middleware(
 instrumentator = Instrumentator()
 instrumentator.instrument(app)
 
+users_db.Base.metadata.create_all(bind=engine)
+tools_db.Base.metadata.create_all(bind=engine)
+rentals_db.Base.metadata.create_all(bind=engine)
+reviews_db.Base.metadata.create_all(bind=engine)
 
-models.Base.metadata.create_all(bind=engine)
+
 
 origins = [
     "http://localhost",
