+++ backend_rentelektro/reviews/handlers.py
@@ -1,6 +1,6 @@
 from sqlalchemy.orm import Session
-from core.models import Review
-from core.schemas.reviews import ReviewCreate
+from reviews.models import Review
+from reviews.schemas import ReviewCreate
 
 def create_review(db: Session, review: ReviewCreate):
     db_review = Review(**review.dict())
